chore: organize files better
