+++ src/components/header/header.test.tsx
@@ -1,28 +1,11 @@
 import React from 'react';
 import { describe, it, expect, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { Header } from '../header';
+import { Header } from './header';
 
 // Mock the ModeToggle component
 vi.mock('../mode-toggle', () => ({
   ModeToggle: () => <div data-testid='mode-toggle'>Mode Toggle</div>,
-}));
-
-// Mock the Logo component
-vi.mock('../logo', () => ({
-  Logo: () => (
-    <div className='flex items-center'>
-      <span className='text-white font-bold text-xl'>Emoji Map</span>
-    </div>
-  ),
-}));
-
-// Mock the Clerk components
-vi.mock('@clerk/nextjs', () => ({
-  SignedIn: ({ children }: React.PropsWithChildren<unknown>) => (
-    <div data-testid='signed-in'>{children}</div>
-  ),
-  UserButton: () => <div data-testid='user-button'>User Button</div>,
 }));
 
 // Mock next-themes
@@ -53,13 +36,6 @@ describe('Header', () => {
     render(<Header />);
 
     expect(screen.getByTestId('mode-toggle')).toBeInTheDocument();
-  });
-
-  it('should render the user button inside SignedIn component', () => {
-    render(<Header />);
-
-    expect(screen.getByTestId('signed-in')).toBeInTheDocument();
-    expect(screen.getByTestId('user-button')).toBeInTheDocument();
   });
 
   it('should have the correct styling', () => {
