+++ src/components/mode-toggle/mode-toggle.tsx
@@ -28,6 +28,7 @@ export function ModeToggle({ className }: { className?: string }) {
           variant='ghost'
           size='icon'
           className={cn(className, 'hover:bg-primary/10')}
+          data-testid='mode-toggle'
         >
           <SunIcon className='h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0 text-foreground' />
           <MoonIcon className='absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100 text-foreground' />
